+++ oligopool/base/core_pack.py
@@ -1438,7 +1438,7 @@ def pack_aggregator(
         filepath=mpath)
 
     # Clear Line
-    liner.send(' ')
+    liner.send(' \r')
 
     # Enqueue Pack Path and Done!
     packqueue.put(mpath)